Move Snapshot capabilities to EE
