+++ images/sds-local-volume-csi/src/driver/controller_snapshots.go
@@ -1,0 +1,25 @@
+//go:build !ee
+
+package driver
+
+import (
+	"context"
+
+	"github.com/container-storage-interface/spec/lib/go/csi"
+)
+
+func (d *Driver) CreateSnapshot(ctx context.Context, request *csi.CreateSnapshotRequest) (*csi.CreateSnapshotResponse, error) {
+	d.log.Info("call method CreateSnapshot")
+	return nil, nil
+
+}
+
+func (d *Driver) DeleteSnapshot(ctx context.Context, request *csi.DeleteSnapshotRequest) (*csi.DeleteSnapshotResponse, error) {
+	d.log.Info("call method DeleteSnapshot")
+	return nil, nil
+}
+
+func (d *Driver) ListSnapshots(_ context.Context, _ *csi.ListSnapshotsRequest) (*csi.ListSnapshotsResponse, error) {
+	d.log.Info("call method ListSnapshots")
+	return nil, nil
+}
